requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
apache-beam==2.65.0
asciitree==0.3.3
asttokens==3.0.0
attrs==25.3.0
backports.tarfile==1.2.0
bokeh==3.7.3
cachetools==5.5.2
Cartopy==0.24.1
certifi==2025.4.26
cffi==1.17.1
cftime==1.6.4.post1
charset-normalizer==3.4.2
chex==0.1.89
click==8.2.0
cloudpickle==3.1.1
colabtools==0.0.1
comm==0.2.2
contourpy==1.3.2
crc32c==2.7.1
crcmod==1.7
cryptography==44.0.3
cycler==0.12.1
dask==2025.4.1
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
dill==0.3.1.1
dinosaur @ git+https://github.com/neuralgcm/dinosaur.git@26c592e6156ccf3bbb4790c5d019cc1f10a99a4a
dinosaur-dycore==1.2.1
dm-haiku==0.0.14
dm-tree==0.1.9
dnspython==2.7.0
docopt==0.6.2
docstring_parser==0.16
donfig==0.8.1.post1
executing==2.2.0
fastavro==1.10.0
fasteners==0.19
flexcache==0.3
flexparser==0.4
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.2
gcsfs==2025.3.2
google-api-core==2.25.0rc1
google-apitools==0.5.31
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.92.0
google-cloud-bigquery==3.32.0
google-cloud-bigquery-storage==2.31.0
google-cloud-bigtable==2.30.1
google-cloud-core==2.4.3
google-cloud-datastore==2.21.0
google-cloud-dlp==3.29.0
google-cloud-language==2.17.1
google-cloud-pubsub==2.29.0
google-cloud-pubsublite==1.12.0
google-cloud-recommendations-ai==0.10.17
google-cloud-resource-manager==1.14.2
google-cloud-spanner==3.54.0
google-cloud-storage==2.19.0
google-cloud-videointelligence==2.16.1
google-cloud-vision==3.10.1
google-crc32c==1.7.1
google-genai==1.14.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphcast @ https://github.com/deepmind/graphcast/archive/master.zip#sha256=1c67e0f5beed196b1a64aa70a5577c8b63f0b9d22410ee3ab40f839f52477e51
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.65.5
grpcio-status==1.65.5
h11==0.16.0
hdfs==2.7.3
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
immutabledict==4.2.1
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jax==0.6.0
jaxlib==0.6.0
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
jmp==0.0.4
joblib==1.5.0
jraph==0.0.6.dev0
jsonpickle==3.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.15
keyring==25.6.0
keyrings.google-artifactregistry-auth==1.1.2
kiwisolver==1.4.8
locket==1.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
ml_dtypes==0.5.1
more-itertools==10.7.0
multidict==6.4.3
mypy_extensions==1.1.0
narwhals==1.39.0
nest-asyncio==1.6.0
numcodecs==0.16.0
numpy==2.2.5
oauth2client==4.1.3
oauthlib==3.2.2
objsize==0.7.1
opentelemetry-api==1.33.0
opentelemetry-sdk==1.33.0
opentelemetry-semantic-conventions==0.54b0
opt_einsum==3.4.0
orjson==3.10.18
overrides==7.7.0
packaging==25.0
pandas==2.2.3
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==11.2.1
Pint==0.24.4
platformdirs==4.3.8
pluggy==1.5.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==16.1.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydot==1.4.2
Pygments==2.19.1
PyJWT==2.9.0
pymongo==4.12.1
pyparsing==3.2.3
pyproj==3.7.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
rechunker==0.5.4
redis==5.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
rsa==4.9.1
rtree==1.4.0
scikit-learn==1.6.1
scipy==1.15.3
SecretStorage==3.3.3
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
stack-data==0.6.3
tabulate==0.9.0
tensorstore==0.1.74
threadpoolctl==3.6.0
toolz==1.0.0
tornado==6.4.2
traitlets==5.14.3
tree-math==0.2.1
trimesh==4.6.8
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
weatherbenchX @ git+https://github.com/google-research/weatherbenchX.git@42a90490e5cf6c9806dc4ff645785a87e6dc9239
websockets==15.0.1
widgetsnbextension==4.0.14
wrapt==1.17.2
xarray==2025.4.0
xarray-beam==0.8.0
xarray-tensorstore==0.1.5
xyzservices==2025.4.0
yarl==1.20.0
zarr==3.0.7
zipp==3.21.0
zstandard==0.23.0
